+++ components/compiler/locals.cpp
@@ -30,7 +30,7 @@ namespace Compiler
         if (iter==collection.end())
             return -1;
 
-        return iter-collection.begin();
+        return static_cast<int>(iter-collection.begin());
     }
 
     bool Locals::search (char type, const std::string& name) const
